compelte

+++ main.c
@@ -11,24 +11,20 @@
 /* ************************************************************************** */
 
 #include "src/codexion.h"
-#include <time.h>
 
-void	*routing(void *argv);
+int main(int argc, char *argv[]) {
+  t_data data;
 
-int	main(int argc, char *argv[])
-{
-	t_data	data;
-
-	if (!ft_parse(argc, argv, &data))
-	{
-		return (0);
-	}
-	if (!init_dongle(&data))
-		destroy_dongle(data.dongle);
-	if (!init_coder(&data))
-		destroy_coder(data.coder);
-	init_thread(&data);
-	join_thread(&data);
-	destroy_all(&data);
-	return (0);
+  if (!ft_parse(argc, argv, &data))
+    return (0);
+  if (!init_dongle(&data))
+    return (destroy_dongle(data.dongle, data.number_of_coders), 0);
+  if (!init_coder(&data))
+    return (destroy_coder(data.coder, data.number_of_coders), 0);
+  if (!init_thread(&data))
+    return (0);
+  if (!join_thread(&data))
+    return (0);
+  destroy_all(&data);
+  return (0);
 }
